backend/src/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
bcrypt
cryptography
pymysql
